Refactor ui lib
