+++ tic-tac-toe/components/Board.js
@@ -1,5 +1,5 @@
 /** @jsx d */
-import { dom as d } from '../../todomvc/lib/state-manager';
+import { dom as d } from '../../todomvc/lib/state-ui-lib';
 import Square from './Square';
 
 // function Board({ squares, onClick }) {
