+++ tic-tac-toe/index.js
@@ -3,7 +3,7 @@
 
 import './index.css';
 import Game from './components/Game';
-import { place } from '../todomvc/lib/state-manager';
+import { place } from '../todomvc/lib/state-ui-lib';
 
 const root = document.getElementById('root');
 place(Game(), root);
